Initial commit
This project provides a wrapped to the EntityFramework ORM. It should
should be easy enough to get NHibernate to work in this project as well
but I have not tried or tested that hypothesis yet. I try to consolidate
the Unit of Work and Repository pattern in a single class as well as
provide other Utility classes like PagedList and some extensions
methods.